Move testing utility functions
